docs(Alert): update alert docs
