+++ src/Alert.js
@@ -5,22 +5,33 @@ import { mapToCssModules, tagPropType } from './utils';
 import Fade from './Fade';
 
 const propTypes = {
+  /** Pass children so this component can wrap the child elements*/
   children: PropTypes.node,
+  /** Add custom class */
   className: PropTypes.string,
+  /** Add custom class for close button */
   closeClassName: PropTypes.string,
+  /** Aria label for close button */
   closeAriaLabel: PropTypes.string,
-  cssModule: PropTypes.object,
+  /** Change color of alert */
   color: PropTypes.string,
+  /** Change existing className with a new className */
+  cssModule: PropTypes.object,
+  /** Toggle fade animation */
   fade: PropTypes.bool,
-  isOpen: PropTypes.bool,
-  toggle: PropTypes.func,
-  tag: tagPropType,
-  transition: PropTypes.shape(Fade.propTypes),
   innerRef: PropTypes.oneOfType([
     PropTypes.object,
     PropTypes.string,
     PropTypes.func,
   ]),
+  /** Control visibility state of Alert */
+  isOpen: PropTypes.bool,
+  /** Set a custom element for this component */
+  tag: tagPropType,
+  /** Function to toggle visibility */
+  toggle: PropTypes.func,
+  /** Props to be passed to `Fade` to modify transition */
+  transition: PropTypes.shape(Fade.propTypes),
 };
 
 const defaultProps = {
