+++ stories/Alert.stories.js
@@ -8,16 +8,17 @@ export default {
         component: `
 [Bootstrap Alert](https://getbootstrap.com/docs/5.1/components/alerts/)
 
-Provide contextual feedback messages for typical user actions.
+Provide contextual feedback messages for typical user actions with the handful of available and flexible alert messages.
         `,
       }
     }
   }
 };
 
-export { default as Alert } from './examples/Alert';
-export { default as AlertContent } from './examples/AlertContent';
-export { default as Dismiss } from './examples/AlertDismiss';
-export { default as AlertLink } from './examples/AlertLink';
-export { default as UncontrolledDismiss } from './examples/AlertUncontrolledDismiss';
-export { default as Props } from './examples/AlertProps';
+export { default as Alert } from './examples/Alert/Alert';
+export { default as Colors } from './examples/Alert/AlertColors';
+export { default as Content } from './examples/Alert/AlertContent';
+export { default as AlertLink } from './examples/Alert/AlertLink';
+export { default as Dismiss } from './examples/Alert/AlertDismiss';
+export { default as UncontrolledDismiss } from './examples/Alert/AlertUncontrolledDismiss';
+export { default as Props } from './examples/Alert/AlertProps';
