+++ stories/examples/Alert/AlertFadeless.js
