+++ stories/examples/Alert/AlertProps.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Alert } from 'reactstrap';
-import Props from './Props';
+import Props from '../Props';
 
 const Example = () => (
   <Props component={Alert} />
