+++ stories/examples/Alert/AlertUncontrolledDismiss.js
@@ -10,3 +10,11 @@ const Example = (props) => {
 }
 
 export default Example;
+
+Example.parameters = {
+  docs: {
+    description: {
+      story: 'And if you are not a fan of using state, you can always use `UncontrolledAlert`.'
+    }
+  },
+}
